Refactor audio and video code to `LLMS_Trait_Audio_Video_Embed`.
